# [BLS] use fast_verify for batch sigs over the same msg