Issue: https://github.com/MystenLabs/fastcrypto/issues/178
Assuming a rogue key protection mechanism has been applied (by proof of possession), BLS batch verify over the same message can use the fastverify function to avoid multiple pairing operations. Interestingly, the multi-threaded mode hasn't been improved, but we see a significant improvement in single-threaded BLST.

- Performance has been improved in single-threaded env -

Before this PR:
```
verification_comparison/BLS12381 single-threaded batched verification/64 (same msg)
            time:  [19.689 ms 19.696 ms 19.704 ms]
```            
After:
```
verification_comparison/BLS12381 single-threaded batched verification/64 (same msg)
            time:  [3.3846 ms 3.3857 ms 3.3870 ms]
```